small changes
made it so that if the bullet does is not hitting an enemy, it will be destroyed. It gets destroyed anyway in the enemy stats script when it collides with an enemy. I did this so I wouldn't have to keep coding if(collision.gameObject.CompareTag(random tag).

turned the block of code that lets you sprint an walk into a function and put it inside the update of player movement script