Finished the Subaru Trophies
Subaru trophies when collected will play their audio clip: "matta ne, bye bye", add +1 to subarus collected, and +25 to the score.

Next thing to work on is a finish game/stage flag. When the flag is touched, it will prompt the gameover screen, but instead of "GAME OVER", it will say "You Win!"

The Subaru above the "GAME OVER" will also say something else that congratulates you on your win.